+++ SystemDotMessaging/Projects/SystemDot.Messaging.Specifications/sequencing/when_receiving_a_message_on_a_durable_channel_that_has_had_sequencing_reset.cs
@@ -3,7 +3,6 @@ using SystemDot.Messaging.Handling;
 using SystemDot.Messaging.Packaging;
 using SystemDot.Messaging.Sequencing;
 using SystemDot.Messaging.Storage;
-using SystemDot.Messaging.Transport.InProcess.Configuration;
 using Machine.Specifications;
 
 namespace SystemDot.Messaging.Specifications.sequencing
@@ -11,11 +10,11 @@ namespace SystemDot.Messaging.Specifications.sequencing
     [Subject(SpecificationGroup.Description)]
     public class when_receiving_a_message_on_a_durable_channel_that_has_had_sequencing_reset : WithMessageConfigurationSubject
     {
-        const int Message = 1;
+        const Int64 Message = 1;
         const string ReceiverAddress = "ReceiverAddress";
         const string SenderAddress = "SenderAddress";
 
-        static TestMessageHandler<int> handler;
+        static TestMessageHandler<Int64> handler;
         static MessagePayload messagePayload;
 
         Establish context = () =>
@@ -31,7 +30,7 @@ namespace SystemDot.Messaging.Specifications.sequencing
             messagePayload.SetSequenceOriginSetOn(DateTime.Today);
             messagePayload.SetFirstSequence(2);
 
-            handler = new TestMessageHandler<int>();
+            handler = new TestMessageHandler<Int64>();
             Resolve<MessageHandlerRouter>().RegisterHandler(handler);
         };
